+++ wakamiti-engine/CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+- Duration data type and Duration assertion.
+
 ### Fixed
 - The junit wakamiti configuration is modified to set the default output path of the current project.
-
-### Added
-- Duration data type
 
 
 ## [2.5.0] - 2024-04-16